feat(cpc): rewrite coupon collection, flavours, and HIP estimation
Rewrite the CpcSketch core to mirror the Apache DataSketches Rust
reference (cpc/{sketch.rs,mod.rs,pair_table.rs}) and wire estimate() to
the ported FM85 HIP/ICON estimator. This replaces the old HLL-style
harmonic-mean dense estimator that produced large errors across flavour
transitions.

- Fields now match the reference: num_coupons, surprising_value_table
  (PairTable), sliding_window, window_offset, first_interesting_column,
  kxp, hip_est_accum, merge_flag.
- Coupon derivation splits one 128-bit xxh3 hash into two 64-bit halves:
  low half gives the row (low lg_k bits), high half gives the column via
  leading-zero count (clipped to 63), matching the reference's two
  64-bit hash outputs. The full-width leading-zero distribution is what
  the HIP/ICON estimator assumes.
- Port row_col_update, update_sparse, promote_sparse_to_windowed,
  update_windowed, move_window, update_hip, refresh_kxp, build_bit_matrix,
  determine_flavor, determine_correct_offset, and PairTable.
- Merge/CpcUnion and to_bytes/from_native_bytes are kept as compiling
  placeholders for Tasks 4 and 5 respectively; documented as such.
- num_coupons() now returns u32; adjusted the cast in serialization.rs.

Real-data accuracy (tests/cpc_accuracy.rs, lg_k=12): 1,000,000 distinct
keys give 0.34% relative error; the cardinality sweep gives 0.31% at
1,000, 1.90% at 50,000, and 0.70% at 500,000, all under the 2%/3%
targets.

+++ src/serialization.rs
@@ -491,7 +491,7 @@ impl Serializable for CpcSketch {
         buf.extend_from_slice(&THETA_SEED_HASH.to_le_bytes()); // bytes 6-7
 
         if !is_empty {
-            buf.extend_from_slice(&(self.num_coupons() as u32).to_le_bytes()); // bytes 8-11
+            buf.extend_from_slice(&self.num_coupons().to_le_bytes()); // bytes 8-11
             // Length of native payload
             buf.extend_from_slice(&(native_bytes.len() as u32).to_le_bytes()); // bytes 12-15
         } else {
